+++ packages/eslint/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tailor-cms/eslint-config
 
+## 2.0.2
+
+### Patch Changes
+
+- Bump dependencies and cleanup runtimes to match Tailor X.
+
 ## 2.0.1
 
 ### Patch Changes